fix: correct inverted profile precedence in am context renderer
The renderer walked profile_layers front-to-back when looking up
winners and building the chain preamble, but Precedence::resolve()
treats profile_layers[i+1] as higher precedence than [i]. With two
profiles defining the same alias, the `from` cell named the loser
and the Shadowed section reversed winner and loser. Snapshot tests
missed it because no two profiles in the fixture shared a name.

Also splits the 1610-line context.rs into a per-section module tree
(escape, preamble, aliases, shadows, variables, invalid,
trust_notice) so tests live next to the code they cover, replaces
the silent `[bug: ...]` fallbacks with `panic!` so invariant
violations surface visibly, threads the new Cow<str> through Row
and the shadow tuples, and rewrites render_preamble as a single
formatdoc template (byte-identical output).
